CONDEC-386: move second test into its own file
